fixed compiling issues

+++ src/smb_trans2.c
@@ -199,7 +199,7 @@ static smb_message  *smb_trans2_find_first (smb_session *s, smb_tid tid, const c
         free(utf_pattern);
         return NULL;
     }
-    msg->packet->header.tid = (uint16_t)tid;
+    msg->packet->header.tid = tid;
 
     SMB_MSG_INIT_PKT(tr2);
     tr2.wct                = 15;
@@ -271,7 +271,7 @@ static smb_message  *smb_trans2_find_next (smb_session *s, smb_tid tid, uint16_t
         free(utf_pattern);
         return NULL;
     }
-    msg_find_next2->packet->header.tid = (uint16_t)tid;
+    msg_find_next2->packet->header.tid = tid;
 
     SMB_MSG_INIT_PKT(tr2_find_next2);
     tr2_find_next2.wct                = 0x0f;
@@ -433,7 +433,7 @@ smb_file  *smb_fstat(smb_session *s, smb_tid tid, const char *path)
         free(utf_path);
         return 0;
     }
-    msg->packet->header.tid = (uint16_t)tid;
+    msg->packet->header.tid = tid;
 
     SMB_MSG_INIT_PKT(tr2);
     tr2.wct                = 15;
